refactor: 리스폰스 부분 분할
