+++ game/src/handlers/index.js
@@ -1,7 +1,4 @@
-import {
-  connectGameRequestHandler,
-  spawnInitialGhostRequestHandler,
-} from './auth/connectGame.handler.js';
+import { spawnInitialGhostRequestHandler } from './auth/connectGame.handler.js';
 import { movePlayerRequestHandler } from './game/player/movePlayer.handler.js';
 import { moveGhostRequestHandler } from './game/ghost/moveGhost.handler.js';
 import { PACKET_TYPE } from '../constants/header.js';
@@ -21,10 +18,6 @@ import { createRoomHandler } from './game/room/createRoom.handler.js';
 import { joinRoomHandler } from './game/room/joinRoom.handler.js';
 
 const handlers = {
-  [PACKET_TYPE.ConnectGameRequest]: {
-    handler: connectGameRequestHandler,
-    protoType: 'common.GamePacket',
-  },
   [PACKET_TYPE.PlayerMoveRequest]: {
     handler: movePlayerRequestHandler,
     protoType: 'common.GamePacket',
